Update compile make targets to support components with library
When compiling a component which makes use of its own component library,
we need to add the component directory to the search paths for
`commodore component compile`.

This commit sets the default value of variable `commodore_args` to

```
--search-paths ./dependencies --search-paths .
````

This variable is used by all compile targets, and some targets now
append `-f tests/$(instance).yml` to the variable instead of setting the
variable from scratch.