Explicitly set `HOME` environment variable for container
This fixes the issues introduced by #54 in `make test` for components
which try to download a Helm chart. With the GID change for the
container in #54, Commodore can't download files to `/app` anymore (the
default home directory of the container), since that directory is owned
by root.

This commit ensures that the Commodore container has a writable home
directory by explicitly setting `HOME` to the component source
directory.