Move test matrix configuration to global variable
Until now, the test matrix has been defined in a local variable for the
`.github/workflows/test.yaml` template. In preparation for generating
Renovate `postUpgradeTasks` and a list of test instances in a make
variable, we need to be able to access the matrix test configuration in
other templates.

This commit moves the test matrix configuration to global variable
`testMatrix`. For now, the GitHub actions template will continue to read
the test matrix from the file-level variable `matrix`, but will fall
back to the global variable `testMatrix` if the file-level variable is
missing.

All already onboarded components should migrate their test matrix (if
any) to global parameter `testMatrix`.