Add make targets `gen-golden-all` and `golden-diff-all`
These targets use a static list of test instances in make variable
`test_instances`. The contents of this variable are generated from the
modulesync parameter `testMatrix.entries`.

We use a generic recursive make target based on the `test_instances`
make variable to implement the new targets. This allows us to reuse the
existing `golden-diff` and `gen-golden` targets.

The new targets are only added for components which have golden and
matrix tests configured and which have migrated their matrix test
configuration to modulesync parameter `testMatrix`.