Add logic to generate Renovate `postUpgradeTasks`
We generate section `postUpgradeTasks` in `renovate.json` for components
which have golden tests enabled. To determine the command to execute, we
check if the component has matrix tests configured (based on parameter
`testMatrix`). We use the `make gen-golden-all` as the command for
components with matrix tests and `make gen-golden` for other components.