Split `COMMODORE_CMD` and `COMPILE_CMD`
We want to be able to adjust which `commodore` to use without having to
reproduce the full compile command when compiling individual components
with `make test`, `make golden-diff` or `make gen-golden`.

This commit reduces `COMMODORE_CMD` to be the equivalent of a bare
invocation of `commodore` and introduces a new variable `COMPILE_COMMAND`
which corresponds to the previous value of `COMMODORE_CMD`.

Matching component-template PR: https://github.com/projectsyn/commodore/pull/437